Support building on Linux.

+++ conanfile.py
@@ -9,13 +9,23 @@ class FreeImageConan(ConanFile):
     package_version = '3'
     version = '%s-%s' % (source_version, package_version)
 
-    requires = 'llvm/3.3-2@vuo/stable'
+    requires = 'llvm/3.3-2@vuo/stable', \
+               'vuoutils/1.0@vuo/stable'
     settings = 'os', 'compiler', 'build_type', 'arch'
     url = 'http://freeimage.sourceforge.net/'
     license = 'http://freeimage.sourceforge.net/license.html'
     description = 'A library to read and write many image formats'
     source_dir = 'FreeImage'
     exports_sources = '*.patch'
+    libs = {
+        'freeimage': 3,
+    }
+
+    def requirements(self):
+        if platform.system() == 'Linux':
+            self.requires('patchelf/0.10pre-1@vuo/stable')
+        elif platform.system() != 'Darwin':
+            raise Exception('Unknown platform "%s"' % platform.system())
 
     def imports(self):
         self.copy('*', '%s/bin' % self.source_dir, 'bin')
@@ -26,10 +36,12 @@ class FreeImageConan(ConanFile):
                   sha256='fbfc65e39b3d4e2cb108c4ffa8c41fd02c07d4d436c594fff8dab1a6d5297f89')
         tools.patch(patch_file='makefile.patch', base_path=self.source_dir)
 
-        self.run('rm -Rf %s/Source/LibJPEG' % self.source_dir)
-        tools.get('http://downloads.sourceforge.net/project/libjpeg-turbo/1.4.2/libjpeg-turbo-1.4.2.tar.gz',
-                  sha256='521bb5d3043e7ac063ce3026d9a59cc2ab2e9636c655a2515af5f4706122233e')
-        self.run('mv libjpeg-turbo-1.4.2 %s/Source/LibJPEG' % self.source_dir)
+        # For now, disable libjpeg-turbo on Linux since it fails with 'undefined symbol: jpeg_resync_to_restart'.
+        if platform.system() == 'Darwin':
+            self.run('rm -Rf %s/Source/LibJPEG' % self.source_dir)
+            tools.get('http://downloads.sourceforge.net/project/libjpeg-turbo/1.4.2/libjpeg-turbo-1.4.2.tar.gz',
+                      sha256='521bb5d3043e7ac063ce3026d9a59cc2ab2e9636c655a2515af5f4706122233e')
+            self.run('mv libjpeg-turbo-1.4.2 %s/Source/LibJPEG' % self.source_dir)
 
         with tools.chdir(self.source_dir):
             self.run('bash gensrclist.sh')
@@ -37,39 +49,61 @@ class FreeImageConan(ConanFile):
         self.run('mv %s/license-fi.txt %s/%s.txt' % (self.source_dir, self.source_dir, self.name))
 
     def build(self):
-        with tools.chdir('%s/Source/LibJPEG' % self.source_dir):
-            autotools = AutoToolsBuildEnvironment(self)
+        import VuoUtils
 
-            # The LLVM/Clang libs get automatically added by the `requires` line,
-            # but this package doesn't need to link with them.
-            autotools.libs = []
+        if platform.system() == 'Darwin':
+            yasm = '/usr/local/bin/yasm'
+        elif platform.system() == 'Linux':
+            yasm = '/usr/bin/yasm'
 
-            autotools.flags.append('-Oz')
+        env_vars = {
+            'CC' : self.deps_cpp_info['llvm'].rootpath + '/bin/clang',
+            'CXX': self.deps_cpp_info['llvm'].rootpath + '/bin/clang++',
+            'NASM': yasm,
+        }
 
-            if platform.system() == 'Darwin':
-                autotools.flags.append('-mmacosx-version-min=10.10')
+        if platform.system() == 'Darwin':
+            with tools.chdir('%s/Source/LibJPEG' % self.source_dir):
+                autotools = AutoToolsBuildEnvironment(self)
 
-            env_vars = {
-                'CC' : self.deps_cpp_info['llvm'].rootpath + '/bin/clang',
-                'CXX': self.deps_cpp_info['llvm'].rootpath + '/bin/clang++',
-                'NASM': '/usr/local/bin/yasm',
-            }
-            with tools.environment_append(env_vars):
-                autotools.configure(build=False,
-                                    host=False,
-                                    args=['--quiet',
-                                          '--disable-shared',
-                                          '--prefix=%s' % os.getcwd()])
-                autotools.make(args=['--quiet'])
+                # The LLVM/Clang libs get automatically added by the `requires` line,
+                # but this package doesn't need to link with them.
+                autotools.libs = []
+
+                autotools.flags.append('-Oz')
+
+                if platform.system() == 'Darwin':
+                    autotools.flags.append('-mmacosx-version-min=10.10')
+                elif platform.system() == 'Linux':
+                    autotools.flags.append('-fPIC')
+
+                with tools.environment_append(env_vars):
+                    autotools.configure(build=False,
+                                        host=False,
+                                        args=['--quiet',
+                                              '--disable-shared',
+                                              '--prefix=%s' % os.getcwd()])
+                    autotools.make(args=['--quiet'])
 
         with tools.chdir(self.source_dir):
-            self.run('make -j9')
-            self.run('mv libfreeimage-%s.dylib-x86_64 libfreeimage.dylib' % self.source_version)
-            self.run('install_name_tool -id @rpath/libfreeimage.dylib libfreeimage.dylib')
+            with tools.environment_append(env_vars):
+                self.run('make -j9')
+
+            if platform.system() == 'Darwin':
+                self.run('mv libfreeimage-%s.dylib-x86_64 libfreeimage.dylib' % self.source_version)
+            elif platform.system() == 'Linux':
+                self.run('mv libfreeimage-%s.so libfreeimage.so' % self.source_version)
+
+            VuoUtils.fixLibs(self.libs, self.deps_cpp_info)
 
     def package(self):
+        if platform.system() == 'Darwin':
+            libext = 'dylib'
+        elif platform.system() == 'Linux':
+            libext = 'so'
+
         self.copy('FreeImage.h', src='%s/Source' % self.source_dir, dst='include')
-        self.copy('*.dylib', src=self.source_dir, dst='lib')
+        self.copy('libfreeimage.%s' % libext, src=self.source_dir, dst='lib')
 
         self.copy('%s.txt' % self.name, src=self.source_dir, dst='license')
